Began work on the ChunkVault implementation.
I made a spatial utility file. Right now all that's included is the find_nearest intersection. This is being used in Chunk and Chunk vault. I'm going to change it so that I can pass in a container instead of iterators.
This required that I add an ::intersectable_type to the Intersection class.

I decided to create a renderable class so that I can pass out Chunk meshes with their associated model matrices. It may make more sense to just change the mesh generation code to take in account the origin.